Move qbb/cbb to avoid cyclic module dependencies
